Style table

+++ db/migrate/20211111064252_create_join_table_recipe_food.rb
@@ -1,7 +1,7 @@
 class CreateJoinTableRecipeFood < ActiveRecord::Migration[6.1]
   def change
     create_join_table :recipes, :foods, table_name: "recipe_foods" do |t|
-      t.decimal :quantity
+      t.integer :quantity
       t.index [:recipe_id, :food_id]
     end
   end
